+++ hw/scripts/openocd.sh
@@ -31,25 +31,25 @@ openocd_load () {
     echo "flash write_image erase $FILE_NAME $FLASH_OFFSET" >> $OCD_CMD_FILE
 
     if [ -z $FILE_NAME ]; then
-	echo "Missing filename"
-	exit 1
+        echo "Missing filename"
+        exit 1
     fi
     if [ ! -f "$FILE_NAME" ]; then
-	# tries stripping current path for readability
+        # tries stripping current path for readability
         FILE=${FILE_NAME##$(pwd)/}
-	echo "Cannot find file" $FILE
-	exit 1
+        echo "Cannot find file" $FILE
+        exit 1
     fi
     if [ -z $FLASH_OFFSET ]; then
-	echo "Missing flash offset"
-	exit 1
+        echo "Missing flash offset"
+        exit 1
     fi
 
     echo "Downloading" $FILE_NAME "to" $FLASH_OFFSET
 
     openocd $CFG -f $OCD_CMD_FILE -c shutdown
     if [ $? -ne 0 ]; then
-	exit 1
+        exit 1
     fi
     rm $OCD_CMD_FILE
     return 0
@@ -62,42 +62,60 @@ openocd_load () {
 openocd_debug () {
     OCD_CMD_FILE=.openocd_cmds
 
+    windows_detect
+
     echo "gdb_port 3333" > $OCD_CMD_FILE
     echo "telnet_port 4444" >> $OCD_CMD_FILE
     echo "$EXTRA_JTAG_CMD" >> $OCD_CMD_FILE
 
     if [ -z "$NO_GDB" ]; then
-	if [ -z $FILE_NAME ]; then
-	    echo "Missing filename"
-	    exit 1
-	fi
-	if [ ! -f "$FILE_NAME" ]; then
-	    echo "Cannot find file" $FILE_NAME
-	    exit 1
-	fi
+        if [ -z $FILE_NAME ]; then
+            echo "Missing filename"
+            exit 1
+        fi
+        if [ ! -f "$FILE_NAME" ]; then
+            echo "Cannot find file" $FILE_NAME
+            exit 1
+        fi
 
-	#
-	# Block Ctrl-C from getting passed to openocd.
-	#
-	set -m
-	openocd $CFG -f $OCD_CMD_FILE -c init -c halt &
-	set +m
+        if [ $WINDOWS -eq 1 ]; then
+            #
+            # Launch openocd in a separate command interpreter, to make sure
+            # it doesn't get killed by Ctrl-C signal from bash.
+            #
+            CFG=`echo $CFG | sed 's/\//\\\\/g'`
+            $COMSPEC "/C start $COMSPEC /C openocd -l openocd.log $CFG -f $OCD_CMD_FILE -c init -c halt"
+        else
+            #
+            # Block Ctrl-C from getting passed to openocd.
+            #
+            set -m
+            openocd $CFG -f $OCD_CMD_FILE -c init -c halt &
+            set +m
+        fi
 
-    	GDB_CMD_FILE=.gdb_cmds
+        GDB_CMD_FILE=.gdb_cmds
 
-    	echo "target remote localhost:3333" > $GDB_CMD_FILE
-	if [ ! -z "$RESET" ]; then
-    	    echo "mon reset halt" >> $GDB_CMD_FILE
+        echo "target remote localhost:3333" > $GDB_CMD_FILE
+        if [ ! -z "$RESET" ]; then
+            echo "mon reset halt" >> $GDB_CMD_FILE
+        fi
+	if [ $WINDOWS -eq 1 ]; then
+	    FILE_NAME=`echo $FILE_NAME | sed 's/\//\\\\/g'`
+            $COMSPEC "/C start $COMSPEC /C arm-none-eabi-gdb -x $GDB_CMD_FILE $FILE_NAME"
+	else
+            arm-none-eabi-gdb -x $GDB_CMD_FILE $FILE_NAME
+            rm $GDB_CMD_FILE
 	fi
-	arm-none-eabi-gdb -x $GDB_CMD_FILE $FILE_NAME
-	rm $GDB_CMD_FILE
     else
-	# No GDB, wait for openocd to exit
-	openocd $CFG -f $OCD_CMD_FILE -c init -c halt
-	return $?
+        # No GDB, wait for openocd to exit
+        openocd $CFG -f $OCD_CMD_FILE -c init -c halt
+        return $?
     fi
 
-    rm $OCD_CMD_FILE
+    if [ ! -x "$COMSPEC" ]; then
+        rm $OCD_CMD_FILE
+    fi
     return 0
 }
